fix: Linuxbrewに対応

+++ install.sh
@@ -19,4 +19,6 @@ do
 done
 
 # Launchpad
+if [ $(uname) = Darwin ]; then 
 defaults write com.apple.dock springboard-columns -int 10;defaults write com.apple.dock springboard-rows -int 6;defaults write com.apple.dock ResetLaunchPad -bool TRUE;killall Dock
+fi
